+++ frontend/src/embedding-sdk-bundle/components/private/SdkQuestion/components/Breakout/use-breakout-data.ts
@@ -7,6 +7,7 @@ import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
 
 import { useSdkQuestionContext } from "../../context";
+import { hasAggregationWithoutBreakoutOnPrevStage } from "../../utils/stages";
 
 export interface SDKBreakoutItem extends BreakoutListItem {
   stageIndex: number;
@@ -85,20 +86,3 @@ export const useBreakoutData = (): SDKBreakoutItem[] => {
         });
     });
 };
-
-/**
- * Matches the notebook editor's logic: a stage is hidden when
- * the previous stage has aggregations but no breakouts.
- * See `getQuestionSteps` in notebook/utils/steps.ts.
- */
-function hasAggregationWithoutBreakoutOnPrevStage(
-  query: Lib.Query,
-  stageIndex: number,
-) {
-  if (stageIndex >= 1) {
-    const hasAggregations = Lib.aggregations(query, stageIndex - 1).length > 0;
-    const hasBreakouts = Lib.breakouts(query, stageIndex - 1).length > 0;
-    return hasAggregations && !hasBreakouts;
-  }
-  return false;
-}
